Finishing touches on wall cross detection
Works well now. If you wander long enough that you get near room walls,
the warning sounds will stop--we need to make sure we instruct users to
back away from sound right away, so this is unlikely to happen. Very
good news is that it seems not to be getting 'stuck' as before.